# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking workouts, exercises, and fitness routines",
        "title": "GetFit API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.yoursite.com/support",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/{userID}/routine": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all workout routines created by a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routines"
                ],
                "summary": "Get all user routines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of routines",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Routine"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to fetch routines",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a structured workout routine with exercise templates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routines"
                ],
                "summary": "Create a new workout routine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Routine information",
                        "name": "routine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.Routine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Routine created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing fields",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to create routine",
                        "schema": {}
                    }
                }
            }
        },
        "/users/{userID}/routine/{routineID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a specific workout routine by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routines"
                ],
                "summary": "Get routine by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Routine ID",
                        "name": "routineID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Routine information",
                        "schema": {
                            "$ref": "#/definitions/store.Routine"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {}
                    },
                    "404": {
                        "description": "Routine not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to fetch routine",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a workout routine from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routines"
                ],
                "summary": "Delete a routine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Routine ID",
                        "name": "routineID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Routine successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to delete routine",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update details of an existing workout routine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routines"
                ],
                "summary": "Update a routine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Routine ID",
                        "name": "routineID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated routine information",
                        "name": "routine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.updateRoutinePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Routine updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing fields",
                        "schema": {}
                    },
                    "404": {
                        "description": "Routine not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to update routine",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.updateRoutinePayload": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "exercise_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "expected_version": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "store.Routine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.RoutineExercise"
                    }
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "store.RoutineExercise": {
            "type": "object",
            "properties": {
                "exercise_id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "template_sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.TemplateSet"
                    }
                }
            }
        },
...
        "store.TemplateSet": {
            "type": "object",
            "properties": {
                "reps": {
                    "type": "integer"
                },
                "set_number": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}